+++ spec/integration/allow_external_requests_spec.rb
@@ -6,8 +6,8 @@ RSpec.describe "miniproxy" do
   let(:session) { Capybara::Session.new(:firefox) }
 
   describe "allowing external requests" do
-    before { MiniProxy::Server.allow_external_requests = true }
-    after { MiniProxy::Server.allow_external_requests = false }
+    before { MiniProxy.allow_external_requests = true }
+    after { MiniProxy.allow_external_requests = false }
 
     it "hits the internet" do
       session.visit("http://example.com")
@@ -16,11 +16,11 @@ RSpec.describe "miniproxy" do
 
     context "when a stub is set" do
       before do
-        MiniProxy::Server.stub_request(method: "GET", url: /example.com/, response: { body: "foo" })
+        MiniProxy.stub_request(method: "GET", url: /example.com/, response: { body: "foo" })
       end
 
       after do
-        MiniProxy::Server.reset
+        MiniProxy.reset
       end
 
       it "uses the stub rather than hitting the internet" do
@@ -32,16 +32,16 @@ RSpec.describe "miniproxy" do
 
     describe "alternating configuration on the fly" do
       it "works as expected when toggling allow_external_requests" do
-        MiniProxy::Server.allow_external_requests = true
+        MiniProxy.allow_external_requests = true
 
         session.visit("http://example.com")
         expect(session).to have_content "Example Domain"
 
-        MiniProxy::Server.allow_external_requests = false
+        MiniProxy.allow_external_requests = false
 
         expect {
           session.visit("http://foo.com")
-          MiniProxy::Server.reset
+          MiniProxy.reset
         }.to output(/WARN/).to_stdout_from_any_process
       end
     end
