+++ spec/lib/miniproxy/server_spec.rb
@@ -1,17 +1,17 @@
-require "miniproxy/server"
+require "miniproxy"
 require "miniproxy/remote"
 
-RSpec.describe MiniProxy::Server do
+RSpec.describe MiniProxy do
   let(:remote_stub) { instance_double(MiniProxy::Remote) }
 
   before do
-    allow(MiniProxy::Server).to receive(:remote).and_return remote_stub
+    allow(MiniProxy).to receive(:remote).and_return remote_stub
   end
 
   describe ".reset" do
     it "calls clear on the remote object" do
       expect(remote_stub).to receive(:clear)
-      MiniProxy::Server.reset
+      MiniProxy.reset
     end
   end
 
@@ -23,7 +23,7 @@ RSpec.describe MiniProxy::Server do
 
       it "calls stop on the remote object" do
         expect(remote_stub).to receive(:stop)
-        MiniProxy::Server.stop
+        MiniProxy.stop
       end
     end
 
@@ -34,7 +34,7 @@ RSpec.describe MiniProxy::Server do
 
       it "does not call stop on the remote object" do
         expect(remote_stub).to_not receive(:stop)
-        MiniProxy::Server.stop
+        MiniProxy.stop
       end
     end
   end
@@ -42,20 +42,20 @@ RSpec.describe MiniProxy::Server do
   describe ".port" do
     it "calls port on the remote object" do
       expect(remote_stub).to receive(:port)
-      MiniProxy::Server.port
+      MiniProxy.port
     end
   end
 
   describe ".host" do
     it "returns '127.0.0.1'" do
-      expect(MiniProxy::Server.host).to eql "127.0.0.1"
+      expect(MiniProxy.host).to eql "127.0.0.1"
     end
   end
 
   describe ".stub_request" do
     it "calls stub_request on the remote object with the given arguments" do
       expect(remote_stub).to receive(:stub_request).with(method: "GET", url: "namco.co.jp", response: { body: "765" })
-      MiniProxy::Server.stub_request(method: "GET", url: "namco.co.jp", response: { body: "765" })
+      MiniProxy.stub_request(method: "GET", url: "namco.co.jp", response: { body: "765" })
     end
   end
 end
